# Conflicts:
#	src/PLAYER/MTS/MonteCarloTreeSearch.java
#	src/PLAYER/MTS/Node.java
#	src/PLAYER/MTS/Tree.java
#	src/PLAYER/MTS/UCT.java
#	src/UI/Grid.java
#	src/UI/Menu.java

+++ src/PLAYER/MTS/ProbabilityNaive.java
@@ -12,7 +12,7 @@ public class ProbabilityNaive {
      * @param numberOfSimulations
      * @return Best value of win rate
      */
-    public static double PNaive(int numberOfWins, int numberOfSimulations) {
+    public static double PNaive(double numberOfWins, int numberOfSimulations) {
 
         if (numberOfSimulations == 0) {return Integer.MAX_VALUE;} // W/N = 2147483647
         double PNaive = (double)(numberOfWins/ numberOfSimulations);
